Ajustando o tipo que estava sendo passado como Long para Integer, onde não podemos utilizar tipos primitivos como parametros

+++ back-end/src/main/java/fatec/morpheus/repository/ApiContentRepository.java
@@ -9,7 +9,7 @@ import fatec.morpheus.entity.Api;
 import fatec.morpheus.entity.ApiContent;
 
 @Repository
-public interface ApiContentRepository extends JpaRepository<ApiContent, Long> {
+public interface ApiContentRepository extends JpaRepository<ApiContent, Integer> {
 
     Optional<ApiContent> findByApiIdAndApiAddressAndDate(Api api, String apiAddress, String date);
 }
